% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecontrol_data.R
\docType{data}
\name{data.list}
\alias{data.list}
\title{Formatted shape and case-control data for Egypt ISIS cases and controls}
\format{
An object of class \code{list} of length 28.
}
\usage{
data.list
}
\description{
Data taken from leaked ISIS border documents and merged with shapefile data.
}
\keyword{datasets}
