% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{Prepare data}
\usage{
data.prep(
  shape,
  survey,
  shape_large.area_id_name = NA,
  shape_large.area_id_num = NA,
  shape_small.area_id_name = NA,
  shape_small.area_id_num = NA,
  survey_small.area_id_num = NA,
  survey_small.area_id_name = NA,
  drop.incomplete.records = NA,
  colnames_X = NA,
  interactions_list = NA,
  scale_X = NA,
  colname_y = NA,
  contamination = T,
  pi = NA,
  large_area_shape = F
)
}
\arguments{
\item{shape}{a shapefile}

\item{survey}{data which can be connected to the shape file by some small-area id;}

\item{shape_large.area_id_name}{large area identifiers in the shapefile;}

\item{shape_large.area_id_num}{large area identifiers in the shapefile;}

\item{shape_small.area_id_name}{small area identifiers in the shapefile;}

\item{survey_small.area_id_num}{small area identifiers in the survey;}

\item{survey_small.area_id_name}{small area identifiers in the survey;}

\item{drop.incomplete.records}{should the function return complete data?;}

\item{colnames_X}{a character vector of covariates definining the design matrix X. Note these should be numeric;}

\item{interactions_list}{a list object where each element is a string of the form "a*b" where a and be are the names of two variables in colnames_X.}

\item{scale_X}{takes values "1sd" or "2sd" corresponding namely to (x-xbar)/sd(x) and (x-xbar)/(2sd(x)). No scaling if NA ;}

\item{colname_y}{variable name for the outcome variable - note this should be numeric ;}

\item{contamination}{the function spits out an offset based on log(P1/P0) (see Rota et al. 2013) - should this offset account for contamination ? ;}

\item{pi}{a scalar defining the prevalence of the outcome in the population of interest ;}

\item{large_area_shape}{hould the function return a large-area shapefile ?}

\item{shape_small.area_id_numsmall}{area identifiers in the shapefile;}
}
\value{

}
\description{
Prepare data in format required for estimation procedure described in "Explaining Recruitment to Violent Extremism: A Bayesian Case-Control Approach.
}
