% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecontrol_data.R
\docType{data}
\name{denom_mena}
\alias{denom_mena}
\title{Denominator data for MENA countries}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9 rows and 7 columns.
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
denom_mena
}
\description{
Male Sunni population over 18 for 9 MENA countries.
}
\details{

}
\keyword{datasets}
