% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgf.R
\name{dgf}
\alias{dgf}
\title{Simulate data generating process}
\usage{
dgf(
  n.sims = 1,
  n = 100,
  pi.hat.naive = 0.5,
  p = 1,
  X_corr = 0,
  pi = 0.05,
  Moran.I.corr = 0.8,
  spatial_structure = "scotland_lipcancer"
)
}
\arguments{
\item{n.sims}{how many samples of simulated data would you like?;}

\item{n}{how large (sample size) should each sample be?;}

\item{pi.hat.naive}{what should be the fraction of cases in the sample?;}

\item{p}{how many normally-distributed covariates should the DGP have?;}

\item{X_corr}{what should be the average correlation among these covariates?;}

\item{pi}{what is the population-level probability of sampling a case?}

\item{Moran.I.corr}{what degree of global spatial autocorrelation (Moran I) should the underlying DGP have?;}

\item{spatial_structure}{on which map should the data be simulated ? (scotland_lipcancer, pennsylvania_lungcancer, and newyork_leukemia)}
}
\value{
# output: a list object with the following variables:

y = see input ;
pi = see input ;
n =  see input ;
p =  see input ;
X =  matrix of simulated covariates, including a vector of 1s for the intercept ;
X_corr = see input ;
small_area_id = see input for 'fit' function ;
beta = true regression coefficient values ;
gamma = true small-area effects ;
mu = true latent propensity of recruitment ;
moranI.mu = empirical Moran I at the individual-level for true latent propensity of recruitment ;
moranI.gamma = empirical Moran I for true small-area effects
}
\description{
Function to simulate data from a relevant data generating process (DGP) ;
currently this function supports the creation of DGPs with 1 layer of area effects (i.e. small-area effects)
}
