% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Function to estimate a Bayesian case-control model. Note this function uses
STAN and the working environment needs to be optimized for STAN fitting
purposes in order to efficiently run the model.}
\usage{
fit(
  data = data_sim,
  contamination = F,
  offset = F,
  beta_prior = "normal",
  small_area_prior = "fixed",
  large_area_prior = NA,
  intercept_scale = NA,
  iter = 100,
  warmup = 50,
  thin = 4,
  cores = 4,
  chains = 4,
  control = list(max_treedepth = 7, adapt_delta = 0.8),
  verbose = TRUE
)
}
\arguments{
\item{data}{data is a list object containing ad-minimum the following elements:}

\item{contamination}{should the model include the rota et al. (2013) style contamination layer?;}

\item{offset}{should the model include a King and Zeng (2001) style offset ? if contamination is also specified, this will be a contaminated-offset ;}

\item{beta_prior}{what prior should the regression coefficients have ? choice between normal and cauchy;}

\item{small_area_prior}{what should be the small-area effects type ? choice between area fixed effects, area random effects, area ICAR effects or area BYM2 effects or NA = no area effects;}

\item{large_area_prior}{what should be the small-area effects type ? choice between area fixed effects, area random effects, or NA = no area effects;}

\item{intercept_scale}{arguments from stan, see \code{?stan} for further information.}

\item{iter}{number of draws; arguments from stan, see \code{?stan} for further information}

\item{warmup}{number of warmup draws}

\item{thin}{number to thin draws}

\item{cores}{numvber of cores to use}

\item{chains}{number of chains}

\item{control}{list; tree depth and adapta delta}

\item{verbose}{print full stan output}

\item{show_code}{should the stan code print on the screen at the end of the run ? (useful for checks);}
}
\value{

}
\description{
Function to estimate a Bayesian case-control model. Note this function uses
STAN and the working environment needs to be optimized for STAN fitting
purposes in order to efficiently run the model.
}
