% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecontrol_data.R
\docType{data}
\name{mon_egypt}
\alias{mon_egypt}
\title{Monitor object from \code{rstan::mon()} of class "simsummary" "data.frame"}
\format{
An object of class \code{simsummary} (inherits from \code{data.frame}) with 2994 rows and 22 columns.
}
\usage{
mon_egypt
}
\description{
Data from generated fit object output of \code{fit()} estimation procedure
}
\keyword{datasets}
