% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{scotland}
\alias{scotland}
\title{Shapefile data for \code{dgf()} function: scotland}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{https://cran.r-project.org/web/packages/SpatialEpi/index.html}
}
\usage{
scotland
}
\description{
Data taken from \code{SpatialEpi} package.
}
\details{

}
\keyword{datasets}
