% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_data.R
\docType{data}
\name{shape_egypt}
\alias{shape_egypt}
\title{Shapefile data for Egypt}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 365 rows and 20 columns.
}
\source{
\url{https://data.humdata.org/dataset/geoboundaries-admin-boundaries-for-egypt/resource/920c7328-f6d1-4882-86f7-45e283cee521}
}
\usage{
shape_egypt
}
\description{
District-level shapefile data for Egypt
}
\details{

}
\keyword{datasets}
