% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecontrol_data.R
\docType{data}
\name{survey_egypt}
\alias{survey_egypt}
\title{Case-control data for Egypt ISIS cases and controls}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 727 rows and 16 columns.
}
\usage{
survey_egypt
}
\description{
Data taken from leaked ISIS border documents..
}
\details{
Includes information on home district, individual-level, and area-level
covariates.
}
\keyword{datasets}
