% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected.R
\name{testconnected}
\alias{testconnected}
\title{Test graph connected}
\usage{
testconnected(nb_object)
}
\arguments{
\item{nb_object}{nb object in prepared data list}
}
\value{

}
\description{
Test if nb object is fully connected. Created by Mitzi Morris.
@seealso \url{https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R}
}
