% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{Prepare data}
\usage{
data.prep(
  shape,
  survey,
  shape_large.area_id_name = NA,
  shape_large.area_id_num = NA,
  shape_small.area_id_name = NA,
  shape_small.area_id_num = NA,
  survey_small.area_id_num = NA,
  survey_small.area_id_name = NA,
  drop.incomplete.records = NA,
  colnames_X = NA,
  interactions_list = NA,
  scale_X = NA,
  colname_y = NA,
  contamination = T,
  pi = NA,
  large_area_shape = F
)
}
\arguments{
\item{shape}{sf object: shapefile data.}

\item{survey}{data.table data.frame, case-control data including common geographic ID.}

\item{shape_large.area_id_name}{string, large area name identifiers in the shapefile.}

\item{shape_large.area_id_num}{integer, large area identifiers in the shapefile.}

\item{shape_small.area_id_name}{string, small area name identifiers in the shapefile.}

\item{shape_small.area_id_num}{integer, small area identifiers in the shapefile.}

\item{survey_small.area_id_num}{string, small area name identifiers in the survey.}

\item{survey_small.area_id_name}{integer, small area identifiers in the survey.}

\item{drop.incomplete.records}{logical, should the function return complete data? Defaults to \code{TRUE}.}

\item{colnames_X}{character vector, covariates definining the design matrix X. Must be numeric.}

\item{interactions_list}{list, each element is a string of the form "a*b" where a and be are the names of two variables in colnames_X.}

\item{scale_X}{string, takes values "1sd" or "2sd."}

\item{colname_y}{string, variable name for the outcome variable. Must be numeric.}

\item{contamination}{logical, should this offset account for contamination? Defaults to \code{TRUE}.}

\item{pi}{numeric, scalar defining the prevalence of the outcome in the population of interest.}

\item{large_area_shape}{logical, should the function return a large-area shapefile? Defaults to \code{TRUE}.}
}
\value{

}
\description{
Prepare data in format required for estimation procedure described in "Explaining Recruitment to Violent Extremism: A Bayesian Case-Control Approach.
}
